// src/Rest/model/tblPaciente.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TblPersona } from './tblPersona';

export interface TblPaciente { 
    estadoPaciente?: boolean;
    idPaciente?: number;
    pacDiscapacidad?: string;
    tblPersona?: TblPersona;
}