// src/Rest/model/tblPersona.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TblPersona { 
    idPersona?: number;
    perApellido?: string;
    perCedula?: string;
    perDireccion?: string;
    perEmail?: string;
    perFechaNac?: string;
    perNombre?: string;
    perTelefono?: string;
}